+++ Analytics/Sources/Delivery/Analytics+Bot.swift
@@ -93,4 +93,8 @@ public extension Analytics {
         let params = ["strategy": strategyName]
         service.track(event: .did, element: .bot, name: "changeHomeFeedStrategy", params: params)
     }
+    
+    func trackBotDidOptimizeSQLite() {
+        service.track(event: .did, element: .bot, name: "optimizeSQLite")
+    }
 }