+++ Source/GoBot/ViewDatabase.swift
@@ -55,153 +55,137 @@ class ViewDatabase {
 
     // TODO: use this to trigger fill on update and wipe previous versions
     // https://app.asana.com/0/914798787098068/1151842364054322/f
-    static var schemaVersion: UInt = 20
+    static let schemaVersion: UInt = 20
 
     // should be changed on login/logout
     private var currentUserID: Int64 = -1
-    private var currentUser: Identity?
+    private(set) var currentUser: Identity?
 
     // skip messages older than this (6 months)
     // this should be removed once the database was refactored
     private var temporaryMessageExpireDate: Double = -60 * 60 * 24 * 30 * 6
     
     // MARK: Tables and fields
-    private let colID = Expression<Int64>("id")
+    let colID = Expression<Int64>("id")
     
-    private let authors: Table
-    private let colAuthor = Expression<Identity>("author")
+    let authors = Table(ViewDatabaseTableNames.authors.rawValue)
+    let colAuthor = Expression<Identity>("author")
     
-    private var msgKeys: Table
-    private let colKey = Expression<MessageIdentifier>("key")
-    private let colHashedKey = Expression<String>("hashed")
+    let msgKeys = Table(ViewDatabaseTableNames.messagekeys.rawValue)
+    let colKey = Expression<MessageIdentifier>("key")
+    let colHashedKey = Expression<String>("hashed")
     
-    private var msgs: Table
-    private let colHidden = Expression<Bool>("hidden")
-    private let colRXseq = Expression<Int64>("rx_seq")
-    private let colMessageID = Expression<Int64>("msg_id")
-    private let colAuthorID = Expression<Int64>("author_id")
-    private let colSequence = Expression<Int>("sequence")
-    private let colMsgType = Expression<String>("type")
+    let msgs = Table(ViewDatabaseTableNames.messages.rawValue)
+    let colHidden = Expression<Bool>("hidden")
+    let colRXseq = Expression<Int64>("rx_seq")
+    let colMessageID = Expression<Int64>("msg_id")
+    let colAuthorID = Expression<Int64>("author_id")
+    let colSequence = Expression<Int>("sequence")
+    let colMsgType = Expression<String>("type")
     // received time in milliseconds since the unix epoch
-    private let colReceivedAt = Expression<Double>("received_at")
-    private let colWrittenAt = Expression<Double>("written_at")
-    private let colClaimedAt = Expression<Double>("claimed_at")
-    private let colDecrypted = Expression<Bool>("is_decrypted")
+    let colReceivedAt = Expression<Double>("received_at")
+    let colWrittenAt = Expression<Double>("written_at")
+    let colClaimedAt = Expression<Double>("claimed_at")
+    let colDecrypted = Expression<Bool>("is_decrypted")
+    /// A flag that is true if this message is not from a real feed. The `WelcomeService` inserts "fake" messages
+    /// into SQLite to help with onboarding.
+    let colOffChain = Expression<Bool>("off_chain")
     
-    private let colMessageRef = Expression<Int64>("msg_ref")
+    let colMessageRef = Expression<Int64>("msg_ref")
 
-    private var abouts: Table
-    private let colAboutID = Expression<Int64>("about_id")
-    private let colName = Expression<String?>("name")
-    private let colImage = Expression<BlobIdentifier?>("image")
-    private let colDescr = Expression<String?>("description")
-    private let colPublicWebHosting = Expression<Bool?>("publicWebHosting")
+    let abouts = Table(ViewDatabaseTableNames.abouts.rawValue)
+    let colAboutID = Expression<Int64>("about_id")
+    let colName = Expression<String?>("name")
+    let colImage = Expression<BlobIdentifier?>("image")
+    let colDescr = Expression<String?>("description")
+    let colPublicWebHosting = Expression<Bool?>("publicWebHosting")
     
-    private var currentBannedContent: Table
+    let currentBannedContent = Table(ViewDatabaseTableNames.bannedContent.rawValue)
     // colID
-    private var colIDType = Expression<Int>("type")
+    let colIDType = Expression<Int>("type")
     
-    private var contacts: Table
+    let contacts = Table(ViewDatabaseTableNames.contacts.rawValue)
     // colAuthorID
-    private let colContactID = Expression<Int64>("contact_id")
-    private let colContactState = Expression<Int>("state")
+    let colContactID = Expression<Int64>("contact_id")
+    let colContactState = Expression<Int>("state")
     
-    private var posts: Table
-    private let colText = Expression<String>("text")
-    private let colIsRoot = Expression<Bool>("is_root")
+    let posts = Table(ViewDatabaseTableNames.posts.rawValue)
+    let colText = Expression<String>("text")
+    let colIsRoot = Expression<Bool>("is_root")
     
-    private var post_blobs: Table
+    let post_blobs = Table(ViewDatabaseTableNames.postBlobs.rawValue)
     // msg_ref
-    private let colIdentifier = Expression<String>("identifier") // TODO: blobHash:ID
+    let colIdentifier = Expression<String>("identifier") // TODO: blobHash:ID
     // colName
-    private let colMetaBytes = Expression<Int?>("meta_bytes")
-    private let colMetaWidth = Expression<Int?>("meta_widht")
-    private let colMetaHeight = Expression<Int?>("meta_height")
-    private let colMetaMimeType = Expression<String?>("meta_mime_type")
-    private let colMetaAverageColorRGB = Expression<Int?>("meta_average_color_rgb")
+    let colMetaBytes = Expression<Int?>("meta_bytes")
+    let colMetaWidth = Expression<Int?>("meta_widht")
+    let colMetaHeight = Expression<Int?>("meta_height")
+    let colMetaMimeType = Expression<String?>("meta_mime_type")
+    let colMetaAverageColorRGB = Expression<Int?>("meta_average_color_rgb")
 
-    private var mentions_feed: Table
+    let mentions_feed = Table(ViewDatabaseTableNames.mentionsFeed.rawValue)
     // msg_ref
-    private let colFeedID = Expression<Int64>("feed_id")
+    let colFeedID = Expression<Int64>("feed_id")
     
-    private var mentions_msg: Table
+    let mentions_msg = Table(ViewDatabaseTableNames.mentionsMsg.rawValue)
     // msg_ref
     // link_id
     
-    private var mentions_image: Table
+    let mentions_image = Table(ViewDatabaseTableNames.mentionsImage.rawValue)
     // msg_ref
     // link_id
     
-    private var votes: Table
-    private let colLinkID = Expression<Int64>("link_id")
-    private let colValue = Expression<Int>("value")
-    private let colExpression = Expression<String?>("expression")
+    let votes = Table(ViewDatabaseTableNames.votes.rawValue)
+    let colLinkID = Expression<Int64>("link_id")
+    let colValue = Expression<Int>("value")
+    let colExpression = Expression<String?>("expression")
 
-    private var channels: Table
+    let channels = Table(ViewDatabaseTableNames.channels.rawValue)
     // id
     // name
-    private let colLegacy = Expression<Bool>("legacy")
+    let colLegacy = Expression<Bool>("legacy")
     
-    private var channelAssigned: Table // what messages are in a channel?
+    // what messages are in a channel?
+    let channelAssigned = Table(ViewDatabaseTableNames.channelsAssigned.rawValue)
     // msg_ref
-    private let colChanRef = Expression<Int64>("chan_ref")
+    let colChanRef = Expression<Int64>("chan_ref")
 
-    private var tangles: Table
-    private let colRoot = Expression<Int64>("root")
+    let tangles = Table(ViewDatabaseTableNames.tangles.rawValue)
+    let colRoot = Expression<Int64>("root")
     
-    private var branches: Table
-    private let colTangleID = Expression<Int64>("tangle_id")
-    private let colBranch = Expression<Int64>("branch")
+    let branches = Table(ViewDatabaseTableNames.branches.rawValue)
+    let colTangleID = Expression<Int64>("tangle_id")
+    let colBranch = Expression<Int64>("branch")
     
-    private var privateRecps: Table
+    let privateRecps = Table(ViewDatabaseTableNames.privateRecps.rawValue)
     // msg_ref
     // contact_id
     
-    private let addresses: Table
-    private let colAddressID = Expression<Int64>("address_id")
+    let addresses: Table = Table(ViewDatabaseTableNames.addresses.rawValue)
+    let colAddressID = Expression<Int64>("address_id")
     // colAboutID
-    private let colAddress = Expression<String>("address")
-    private let colWorkedLast = Expression<Date?>("worked_last")
-    private let colLastErr = Expression<String>("last_err")
-    private let colUse = Expression<Bool>("use")
-    private let colRedeemed = Expression<Double?>("redeemed")
+    let colAddress = Expression<String>("address")
+    let colWorkedLast = Expression<Date?>("worked_last")
+    let colLastErr = Expression<String>("last_err")
+    let colUse = Expression<Bool>("use")
+    let colRedeemed = Expression<Double?>("redeemed")
     
     // Reports
-    private var reports: Table
+    let reports = Table(ViewDatabaseTableNames.reports.rawValue)
     // colMessageRef
     // colAuthorID
-    private let colReportType = Expression<String>("type")
-    private let colCreatedAt = Expression<Double>("created_at")
+    let colReportType = Expression<String>("type")
+    let colCreatedAt = Expression<Double>("created_at")
     
     // Pubs
-    private var pubs: Table
+    let pubs = Table(ViewDatabaseTableNames.pubs.rawValue)
     // colMessageRef
-    private let colHost = Expression<String>("host")
-    private let colPort = Expression<Int>("port")
+    let colHost = Expression<String>("host")
+    let colPort = Expression<Int>("port")
     // colKey
-
-    init() {
-        self.addresses = Table(ViewDatabaseTableNames.addresses.rawValue)
-        self.authors = Table(ViewDatabaseTableNames.authors.rawValue)
-        self.msgKeys = Table(ViewDatabaseTableNames.messagekeys.rawValue)
-        self.msgs = Table(ViewDatabaseTableNames.messages.rawValue)
-        self.abouts = Table(ViewDatabaseTableNames.abouts.rawValue)
-        self.channels = Table(ViewDatabaseTableNames.channels.rawValue)
-        self.channelAssigned = Table(ViewDatabaseTableNames.channelsAssigned.rawValue)
-        self.contacts = Table(ViewDatabaseTableNames.contacts.rawValue)
-        self.currentBannedContent = Table(ViewDatabaseTableNames.bannedContent.rawValue)
-        self.privateRecps = Table(ViewDatabaseTableNames.privateRecps.rawValue)
-        self.posts = Table(ViewDatabaseTableNames.posts.rawValue)
-        self.post_blobs = Table(ViewDatabaseTableNames.postBlobs.rawValue)
-        self.mentions_msg = Table(ViewDatabaseTableNames.mentionsMsg.rawValue)
-        self.mentions_feed = Table(ViewDatabaseTableNames.mentionsFeed.rawValue)
-        self.mentions_image = Table(ViewDatabaseTableNames.mentionsImage.rawValue)
-        self.votes = Table(ViewDatabaseTableNames.votes.rawValue)
-        self.tangles = Table(ViewDatabaseTableNames.tangles.rawValue)
-        self.branches = Table(ViewDatabaseTableNames.branches.rawValue)
-        self.reports = Table(ViewDatabaseTableNames.reports.rawValue)
-        self.pubs = Table(ViewDatabaseTableNames.pubs.rawValue)
-    }
+    
+    // Search
+    private let postSearch = VirtualTable("post_search")
 
     // MARK: open / close / stats
 
@@ -249,6 +233,36 @@ class ViewDatabase {
                 try db.execute("ALTER TABLE `blocked_content` RENAME TO `banned_content`;")
                 db.userVersion = 10
             }
+            if db.userVersion == 10 {
+                // Run this as a migration since we haven't ever run it before.
+                try db.execute("PRAGMA optimize;")
+                db.userVersion = 11
+            }
+            if db.userVersion == 11 {
+                try db.run(
+                    postSearch.create(
+                        .FTS4(
+                            FTS4Config()
+                                .column(colMessageRef)
+                                .column(colText)
+                        )
+                    )
+                )
+                let posts = try db.prepare(posts)
+                for post in posts {
+                    try db.run(
+                        postSearch.insert(
+                            colMessageRef <- post[colMessageRef],
+                            colText <- post[colText]
+                        )
+                    )
+                }
+                db.userVersion = 12
+            }
+            if db.userVersion == 12 {
+                try db.execute("ALTER TABLE messages ADD off_chain INTEGER NOT NULL DEFAULT (0);")
+                db.userVersion = 13
+            }
         }
     }
     
@@ -265,6 +279,19 @@ class ViewDatabase {
     }
     
     func close() {
+        if let db = openDB {
+            do {
+                try db.execute("PRAGMA analysis_limit = 400;")
+                try db.execute("PRAGMA optimize;")
+                // this event is mostly so that I can make sure this works reliably in production and the app
+                // isn't being terminated before it's done.
+                Analytics.shared.trackBotDidOptimizeSQLite()
+                Log.info("Finished optimizing db")
+            } catch {
+                Log.optional(error)
+                CrashReporting.shared.reportIfNeeded(error: error)
+            }
+        }
         self.openDB = nil
         self.currentUser = nil
         self.currentUserID = -1
@@ -735,6 +762,31 @@ class ViewDatabase {
         return abouts
     }
     
+    func abouts(withNameLike queryString: String) throws -> [About] {
+        guard let db = self.openDB else {
+            throw ViewDatabaseError.notOpen
+        }
+        
+        let query = self.abouts
+            .join(self.authors, on: colID == self.abouts[colAboutID])
+            .order(self.abouts[colName])
+            .where(colName.like("%\(queryString)%"))
+        
+        var abouts: [About] = []
+        
+        let aboutsQuery = try db.prepare(query)
+        for about in aboutsQuery {
+            let about = About(
+                about: try about.get(colAuthor),
+                name: try about.get(colName),
+                description: try about.get(colDescr),
+                imageLink: try about.get(colImage)
+            )
+            abouts += [about]
+        }
+        return abouts
+    }
+    
     // MARK: follows and blocks
     
     // who is this feed following?
@@ -1021,10 +1073,7 @@ class ViewDatabase {
 
     // MARK: recent
     func recentPosts(strategy: FeedStrategy, limit: Int, offset: Int? = nil) throws -> KeyValues {
-        guard let connection = self.openDB else {
-            throw ViewDatabaseError.notOpen
-        }
-        return try strategy.fetchKeyValues(connection: connection, userId: currentUserID, limit: limit, offset: offset)
+        return try strategy.fetchKeyValues(database: self, userId: currentUserID, limit: limit, offset: offset)
     }
 
     func numberOfRecentPosts(with strategy: FeedStrategy, since message: MessageIdentifier) throws -> Int {
@@ -1111,98 +1160,19 @@ class ViewDatabase {
     }
     // table.filter(!(array.contains(id)))
 
-    private func mapQueryToKeyValue(qry: Table) throws -> [KeyValue] {
+    private func mapQueryToKeyValue(qry: Table, useNamespacedTables: Bool = false) throws -> [KeyValue] {
         guard let db = self.openDB else {
             throw ViewDatabaseError.notOpen
         }
-        let colRootMaybe = Expression<Int64?>("root")
-
-        // TODO: add switch over type (to support contact, vote, gathering, etc..)
 
         return try db.prepare(qry).compactMap { row in
-            // tried 'return try row.decode()'
-            // but failed - see https://github.com/VerseApp/ios/issues/29
-            
-            let msgID = try row.get(colMessageID)
-            
-            let msgKey = try row.get(colKey)
-            let msgAuthor = try row.get(colAuthor)
-
-            var c: Content
-               
-            let type = try row.get(self.msgs[colMsgType])
-            
-            switch type {
-            case ContentType.post.rawValue:
-                
-                var rootKey: Identifier?
-                if let rootID = try row.get(colRootMaybe) {
-                    rootKey = try self.msgKey(id: rootID)
-                }
-                
-                let p = Post(
-                    blobs: try self.loadBlobs(for: msgID),
-                    mentions: try self.loadMentions(for: msgID),
-                    root: rootKey,
-                    text: try row.get(colText)
-                )
-                
-                c = Content(from: p)
-                
-            case ContentType.vote.rawValue:
-                
-                let lnkID = try row.get(colLinkID)
-                let lnkKey = try self.msgKey(id: lnkID)
-                
-                let rootID = try row.get(colRoot)
-                let rootKey = try self.msgKey(id: rootID)
-                
-                let cv = ContentVote(
-                    link: lnkKey,
-                    value: try row.get(colValue),
-                    expression: try row.get(colExpression),
-                    root: rootKey,
-                    branches: [] // TODO: branches for root
-                )
-                
-                c = Content(from: cv)
-            case ContentType.contact.rawValue:
-                if let state = try? row.get(colContactState) {
-                    let following = state == 1
-                    let cc = Contact(contact: msgAuthor, following: following)
-                    
-                    c = Content(from: cc)
-                } else {
-                    // Contacts stores only the latest message
-                    // So, an old follow that was later unfollowed won't appear here.
-                    return nil
-                }
-            default:
-                throw ViewDatabaseError.unexpectedContentType(type)
-            }
-            
-            let v = Value(
-                author: msgAuthor,
-                content: c,
-                hash: "sha256", // only currently supported
-                previous: nil, // TODO: .. needed at this level?
-                sequence: try row.get(colSequence),
-                signature: "verified_by_go-ssb",
-                timestamp: try row.get(colClaimedAt)
+            return try KeyValue(
+                row: row,
+                database: self,
+                useNamespacedTables: true,
+                hasMentionColumns: false,
+                hasReplies: false
             )
-            var keyValue = KeyValue(
-                key: msgKey,
-                value: v,
-                timestamp: try row.get(colReceivedAt)
-            )
-            keyValue.metadata.author.about = About(
-                about: msgAuthor,
-                name: try row.get(self.abouts[colName]),
-                description: try row.get(colDescr),
-                imageLink: try row.get(colImage)
-            )
-            keyValue.metadata.isPrivate = try row.get(colDecrypted)
-            return keyValue
         }
     }
     
@@ -1356,7 +1326,7 @@ class ViewDatabase {
             .order(colMessageID.desc)
             .limit(limit)
         
-        return try self.mapQueryToKeyValue(qry: qry)
+        return try self.mapQueryToKeyValue(qry: qry, useNamespacedTables: true)
     }
 
     // PARGMA MARK: - Reports
@@ -1552,31 +1522,38 @@ class ViewDatabase {
         return authorID
     }
     
+    // MARK: - Fetching Posts
+    
     func post(with id: MessageIdentifier) throws -> KeyValue {
         let msgId = try self.msgID(of: id, make: false)
-        // TODO: add 2nd signature to get message by internal ID
-//        guard let db = self.openDB else {
-//            throw ViewDatabaseError.notOpen
-//        }
-//        let msgId = try self.msgID(of: key, make: false)
-//
-//        return self.get(msgID: msgID)
-//    }
-//
-//    func get(msgID: Int64) throws -> KeyValue {
+        return try post(with: msgId)
+    }
+    
+    func post(with messageRef: Int64) throws -> KeyValue {
         guard let db = self.openDB else {
             throw ViewDatabaseError.notOpen
         }
+
+        // TODO: add 2nd signature to get message by internal ID
+        // guard let db = self.openDB else {
+        //         throw ViewDatabaseError.notOpen
+        //     }
+        //     let msgId = try self.msgID(of: key, make: false)
+        //
+        //     return self.get(msgID: msgID)
+        // }
+        //
+        // func get(msgID: Int64) throws -> KeyValue {
         
         let colTypeMaybe = Expression<String?>("type")
         let typeMaybe = try db.scalar(self.msgs
             .select(colTypeMaybe)
-            .filter(colMessageID == msgId)
+            .filter(colMessageID == messageRef)
             .filter(colHidden == false))
 
         guard let msgType = typeMaybe else {
-            Log.unexpected(.botError, "[viewdb] should have type for this message: \(msgId): \(id)")
-            throw ViewDatabaseError.unknownMessage(id)
+            Log.unexpected(.botError, "[viewdb] should have type for this message: \(messageRef)")
+            throw ViewDatabaseError.unknownMessage(String(messageRef))
         }
         
         guard let ct = ContentType(rawValue: msgType) else {
@@ -1591,21 +1568,41 @@ class ViewDatabase {
                     .join(self.posts, on: self.posts[colMessageRef] == self.msgs[colMessageID])
                     .join(.leftOuter, self.tangles, on: self.tangles[colMessageRef] == self.msgs[colMessageID])
                     .join(self.authors, on: self.authors[colID] == self.msgs[colAuthorID])
-                    .join(self.abouts, on: self.abouts[colAboutID] == self.msgs[colAuthorID])
-                    .filter(colMessageID == msgId)
+                    .join(.leftOuter, self.abouts, on: self.abouts[colAboutID] == self.msgs[colAuthorID])
+                    .filter(colMessageID == messageRef)
                     .limit(1)
                 
                 let kv = try self.mapQueryToKeyValue(qry: qry)
                 
                 if kv.count != 1 {
                     Log.unexpected(.botError, "[viewdb] could not find post after we had the type!?")
-                    throw ViewDatabaseError.unknownMessage(id)
+                    throw ViewDatabaseError.unknownMessage(String(messageRef))
                 }
                 return kv[0]
             
         default:
             throw ViewDatabaseError.unhandledContentType(ct)
         }
+    }
+    
+    func posts(matching text: String) throws -> [KeyValue] {
+        guard let connection = self.openDB else {
+            throw ViewDatabaseError.notOpen
+        }
+        
+        // probably need to escape some characters here
+        var messages = [KeyValue]()
+        let query = try connection.prepare(
+            postSearch
+                .filter(colText.match(text))
+        )
+        for row in query {
+            let messageID = row[colMessageRef]
+            let message = try post(with: messageID)
+            messages.append(message)
+        }
+        
+        return messages
     }
     
     // MARK: channels
@@ -1910,12 +1907,21 @@ class ViewDatabase {
             try self.insertPrivateRecps(msgID: msgID, recps: p.recps)
         }
         
-        try db.run(self.posts.insert(
-            colMessageRef <- msgID,
-            colIsRoot <- p.root == nil,
-            colText <- p.text
-        ))
+        try db.run(
+            self.posts.insert(
+                colMessageRef <- msgID,
+                colIsRoot <- p.root == nil,
+                colText <- p.text
+            )
+        )
         
+        try db.run(
+            postSearch.insert(
+                colMessageRef <- msgID,
+                colText <- p.text.lowercased()
+            )
+        )
+
         try self.insertBranches(msgID: msgID, root: p.root, branches: p.branch)
         
         if let m = p.mentions {
@@ -2208,7 +2214,8 @@ class ViewDatabase {
                     colMsgType <- msg.value.content.type.rawValue,
                     colReceivedAt <- msg.timestamp,
                     colClaimedAt <- claimed,
-                    colWrittenAt <- Date().millisecondsSince1970
+                    colWrittenAt <- Date().millisecondsSince1970,
+                    colOffChain <- msg.offChain ?? false
                 ))
             } else {
                 do {
@@ -2220,7 +2227,8 @@ class ViewDatabase {
                         colMsgType <- msg.value.content.type.rawValue,
                         colReceivedAt <- msg.timestamp,
                         colClaimedAt <- claimed,
-                        colWrittenAt <- Date().millisecondsSince1970
+                        colWrittenAt <- Date().millisecondsSince1970,
+                        colOffChain <- msg.offChain ?? false
                     ))
                 } catch Result.error(let errMsg, let errCode, _) {
                     // this is _just_ here because of a fetch-duplication bug in go-ssb
@@ -2348,7 +2356,7 @@ class ViewDatabase {
         ))
     }
 
-    private func msgKey(id: Int64) throws -> MessageIdentifier {
+    func msgKey(id: Int64) throws -> MessageIdentifier {
         guard let db = self.openDB else {
             throw ViewDatabaseError.notOpen
         }
@@ -2517,35 +2525,11 @@ class ViewDatabase {
         }
     }
     
-    private func loadMentions(for msgID: Int64) throws -> [Mention] {
-        guard let db = self.openDB else {
-            throw ViewDatabaseError.notOpen
-        }
-        
+    func loadMentions(for msgID: Int64) throws -> [Mention] {
         // load mentions for this message
-        let feedQry = self.mentions_feed.where(colMessageRef == msgID)
-        let feedMentions: [Mention] = try db.prepare(feedQry).map {
-            row in
-            
-            let feedID = try row.get(colFeedID)
-            let feed = try self.author(from: feedID)
-            
-            return Mention(
-                link: feed,
-                name: try row.get(colName) ?? ""
-            )
-        }
+        let feedMentions = try loadFeedMentions(for: msgID)
+        let msgMentions = try loadMessageMentions(for: msgID)
         
-        let msgMentionQry = self.mentions_msg.where(colMessageRef == msgID)
-        let msgMentions: [Mention] = try db.prepare(msgMentionQry).map {
-            row in
-            
-            let linkID = try row.get(colLinkID)
-            return Mention(
-                link: try self.msgKey(id: linkID),
-                name: ""
-            )
-        }
         /*
         // TODO: We don't =populate mentions_images... so why are we looking it up?
         let imgMentionQry = self.mentions_image
@@ -2564,6 +2548,43 @@ class ViewDatabase {
         */
         
         return feedMentions + msgMentions
+    }
+    
+    func loadFeedMentions(for msgID: Int64) throws -> [Mention] {
+        guard let db = self.openDB else {
+            throw ViewDatabaseError.notOpen
+        }
+
+        let feedQry = mentions_feed.where(colMessageRef == msgID)
+        let feedMentions: [Mention] = try db.prepare(feedQry).map { mentionRow in
+
+            let feedID = try mentionRow.get(colFeedID)
+            let feed = try author(from: feedID)
+
+            return Mention(
+                link: feed,
+                name: try mentionRow.get(colName) ?? ""
+            )
+        }
+
+        return feedMentions
+    }
+
+    func loadMessageMentions(for msgID: Int64) throws -> [Mention] {
+        guard let db = self.openDB else {
+            throw ViewDatabaseError.notOpen
+        }
+
+        let msgMentionQry = mentions_msg.where(colMessageRef == msgID)
+        let msgMentions: [Mention] = try db.prepare(msgMentionQry).map { mentionRow in
+            let linkID = try mentionRow.get(colLinkID)
+            return Mention(
+                link: try msgKey(id: linkID),
+                name: ""
+            )
+        }
+
+        return msgMentions
     }
     
     private func insertBlobs(msgID: Int64, blobs: [Blob]) throws {
@@ -2587,7 +2608,7 @@ class ViewDatabase {
         }
     }
 
-    private func loadBlobs(for msgID: Int64) throws -> [Blob] {
+    func loadBlobs(for msgID: Int64) throws -> [Blob] {
         guard let db = self.openDB else {
             throw ViewDatabaseError.notOpen
         }
@@ -2595,7 +2616,6 @@ class ViewDatabase {
         // let supportedMimeTypes = [MIMEType.jpeg, MIMEType.png]
         
         let qry = self.post_blobs.where(colMessageRef == msgID)
-            .filter(colMetaMimeType == "image/jpeg" || colMetaMimeType == "image/png" )
         
         let blobs: [Blob] = try db.prepare(qry).map {
             row in